// apps/web/src/api/generated/model/onboardingTemplateResponseDto.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * BTN HRMS API
 * The Human Resource Management System API description
 * OpenAPI spec version: 1.0
 */
import type { OnboardingTemplateItemResponseDto } from './onboardingTemplateItemResponseDto';
import type { OnboardingTemplateResponseDtoDepartmentId } from './onboardingTemplateResponseDtoDepartmentId';
import type { OnboardingTemplateResponseDtoPositionId } from './onboardingTemplateResponseDtoPositionId';

export interface OnboardingTemplateResponseDto {
  id: string;
  name: string;
  type: string;
  departmentId?: OnboardingTemplateResponseDtoDepartmentId;
  positionId?: OnboardingTemplateResponseDtoPositionId;
  isDefault: boolean;
  isActive: boolean;
  items: OnboardingTemplateItemResponseDto[];
  createdAt: string;
  updatedAt: string;
}